discuss v0.0.1 Discuss.TopicController

The Topic Controller is the heart of this application. Most of he logic requests are routed here and processed.

Functions

classifyimage(conn, params)

The classifyimage(conn, params) function is used to classify the image. Also, on a sucessful upload of the file, the file is copied to the temp and the file name is modified to avoid the conflict in the file names gnereated by the various users. The file name is relaed with their gihub profile id and the timestamp of sec after the georgian time, this way there will no name collisions. The obtained file is then passed to the CNN which is trained already. The file user has sent is first validated to check if its an image. If the file is an image then the tensorflow code will use the trained weights and clasify the digit.

startindex(conn, params)

The start index functionality lets the users to be redirected to the start index page unless they signin using the github. This page works as a shield to the applcation, where users have to first signin and then they are automatically redicted to the applicaiton page only after a sucessful signon. The users are required to authenticate themselves using their github account.

useapp(conn, params)

Once the user sucessfully login, they are redirected ot his page. The user is provided with the page where they can upload an image for accesing the CNN powered digit classification.

welcome(conn, params)

The welcome page is a simple dummy page with no logic. this is a test page.